Custom Difficulties YOOOO
I know those of you that are watching the github are probably excited because of this commit, but sit tight, there isn't a release yet.  I'm working on getting it to work in Story Mode as well, as far as getting the "Logo"s for the diffs and loading it properly.